GetApis

+++ src/main/java/com/mfs/merchantQR/repo/TblMcRequestRepo.java
@@ -3,6 +3,6 @@ package com.mfs.merchantQR.repo;
 import com.mfs.merchantQR.model.TblMcRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TblMcRequestRepo extends JpaRepository<TblMcRequest, Long> {
-    TblMcRequest findByMcRequestId(Long limitId);
+public interface TblMcRequestRepo extends JpaRepository<TblMcRequest, Integer> {
+    TblMcRequest findByMcRequestId(int limitId);
 }
