+++ src/main/java/com/mfs/merchantQR/service/CommonService.java
@@ -28,7 +28,7 @@ public interface CommonService {
     Response mcAction(McActionRequest mcActionRequest);
 
     //    for getting single MC request object
-    TblMcRequest getMcRequestCheckerById(Long mcRequestId);
+    TblMcRequest getMcRequestCheckerById(int mcRequestId);
 
     String logs(String endPoint, String logLevel, String className, String methodName, String packageDetails, Request request, String message, Response resp) throws HttpClientErrorException;
